// src/pokemons/interfaces/pokemon-response.ts
export interface PokemonsResponse {
  count:    number;
  next:     null;
  previous: null;
  results:  Result[];
}

export interface Result {
  name: string;
  url:  string;
}

export interface SimplePokemon {
  id: string;
  name: string;
  types: string[];
}